Split db_init.cpp into two files
