+++ libraries/chain/database.cpp
@@ -24,11 +24,12 @@
 #include "db_balance.cpp"
 #include "db_block.cpp"
 #include "db_debug.cpp"
+#include "db_genesis.cpp"
 #include "db_getter.cpp"
 #include "db_init.cpp"
 #include "db_maint.cpp"
 #include "db_management.cpp"
 #include "db_market.cpp"
+#include "db_notify.cpp"
 #include "db_update.cpp"
 #include "db_witness_schedule.cpp"
-#include "db_notify.cpp"
